Fix issue 13

+++ prebuild.py
@@ -9,7 +9,7 @@ src = "../../src"
 
 os.makedirs(example_lib_dir, 0o755, True)
 
-if not os.path.exists(dst):
+if not os.path.exists(dst) and os.name != 'nt':
     os.symlink(src, dst)
+    print("Symbolic link created successfully")
 
-print("Symbolic link created successfully")
